Core scripts of AudiImage
RunComputationOfSai.m
RunProcessingOfSai.m
These two run scripts are the main entry points for the computation and
processing of SAI from underlying CARFAC model.

ParametersToComputeSai.m
ParametersToProcessSai.m
Parameter files for configuration of computation and processing of SAI.

script/...
These scripts contain the functions for computation and processing of
SAI and are called by the run scripts.

On branch ForkAudiImage
Changes to be committed:
	new file:   ParametersToComputeSai.m
	new file:   ParametersToProcessSai.m
	new file:   RunComputationOfSai.m
	new file:   RunProcessingOfSai.m
	new file:   script/ComputeSai.m
	new file:   script/PlotAndSavePictureOfCompleteCochleagram.m
	new file:   script/PlotAndSavePictureOfCompletePitchogram.m
	new file:   script/PlotAndSavePictureOfSpectrogram.m
	new file:   script/ProcessFramesForMovieWithComposedImage.m
	new file:   script/ProcessFramesForMovieWithFullfledgedFiguresWithCompleteGrams.m
	new file:   script/ProcessFramesForMovieWithFullfledgedFiguresWithMovieGrams.m
	new file:   script/ProcessSai.m