Modifications to the original code base
matlab/SAI_RunLayered.m
Modifications done are described in detail in the comment header of the
file. But main reason for the modifications is to use the SAI for
analysis and post-processing. The SAI, cochleagram, and pitchogram are
saved as MAT files for further use.

matlab/MakeMovieFromPngsAndWav.m
Modifications done are described in detail in the comment header of the
file. These modifications do not alter functionality; they ensure the
function works and add machine-type independence.

On branch ForkAudiImage
Changes to be committed:
	modified:   matlab/MakeMovieFromPngsAndWav.m
	modified:   matlab/SAI_RunLayered.m