README.md updated with consequences
README.md
Updated to align with the latest project structure and documentation.

docs/README.md
Added a new README file in the docs directory to provide an detailed
documentation on AudiImage. This file contains for now only a
placeholder, documentation work is in progress.

docs/CARFAC_Design_Doc.txt
This is an original file for documentation worth to keep, but moved to
the docs directory for better organization.

pitchogram_demo.jpg
Delete: This file was removed as it was not needed anymore in this fork.

resources/
Add: New resources directory created to hold images and other assets.

On branch ForkAudiImage
Changes to be committed:
	modified:   README.md
	renamed:    CARFAC_Design_Doc.txt -> docs/CARFAC_Design_Doc.txt
	new file:   docs/README.md
	deleted:    pitchogram_demo.jpg
	new file:   resources/AudiImage_GitHubSocialMediaPreview.png
	new file:   resources/START_HERE.png