Double transpose replaced by setting dimension
script/ComputeSai.m
Modify: Helper function ConvertStereoChannelsToMonoChannel() used up to
now in mean() a double transpose of input signal to convert stereo
channels to mono channel. This is replaced by setting the dimension
where mean() should return the mean of the input signal. This "dewarns"
the code and is more efficient.

On branch master
Changes to be committed:
	modified:   script/ComputeSai.m