Load explicitly variables from summary file
script/ProcessSai.m
Modify: Variables needed for processing SAI are now loaded explicitly
from the summary file. This "dewarns" the code, makes it clearer and
consistency with other scripts is improved.
Modify: Folder for temporary frames (and later on the filename of ZIP
file) is now derived from the movie filename to avoid duplication of
code.
Fix: Typo in the head comment.

On branch LoadToStructInProcessSai
Changes to be committed:
	modified:   script/ProcessSai.m