Add remark about AUTHORS and CONTRIBUTORS files
After consulting with an OSS expert based at my company, I have added a
remark about the AUTHORS and CONTRIBUTORS files from the upstream
project to the NOTICE file since both files will remain in this fork.

To be clear, the NOTICE file is from now on the place for authors and
contributors of this fork, while the AUTHORS and CONTRIBUTORS files
are for the upstream project and to respect their work.

On branch AdministrationOssStuff
Changes to be committed:
	modified:   NOTICE

Signed-off-by: Michael Kuka (b0sch-mike) <michael.kuka@de.bosch.com>